Added HTML titles, favicons, and added new css refs
Added HTML site titles to every page along with adding the favicon seen in the index. Also added the names of the new stylesheets that were serperated out for code efficiency reasons. In addition added link to index page when users click the site title in the sidebar
Index: Changed wording of blurb to remove Login references. Added click to enter button and styled it accordingly.